+++ packages/wallet/src/redux/protocols/funding/components/b-wait-for-post-fund-setup.tsx
