+++ packages/wallet/src/redux/protocols/funding/components/funding-error.tsx
